Revert "Hw 26 2"
