+++ src/main/java/com/internet/shop/controller/DeleteUserController.java
@@ -1,4 +1,4 @@
-package com.internet.shop.controller.user;
+package com.internet.shop.controller;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.service.ShoppingCartService;
