+++ src/main/java/com/internet/shop/controller/ProductsController.java
@@ -1,26 +1,24 @@
-package com.internet.shop.controller.user;
+package com.internet.shop.controller;
 
 import com.internet.shop.lib.Injector;
-import com.internet.shop.model.User;
-import com.internet.shop.service.UserService;
+import com.internet.shop.service.ProductService;
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/user/all")
-public class UsersController extends HttpServlet {
+@WebServlet("/product/all")
+public class ProductsController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("com.internet.shop");
-    private UserService userService = (UserService) injector.getInstance(UserService.class);
+    private ProductService productService = (ProductService) injector
+            .getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<User> allUsers = userService.getAll();
-        req.setAttribute("users", allUsers);
-        req.getRequestDispatcher("/WEB-INF/view/user/all.jsp").forward(req, resp);
+        req.setAttribute("products", productService.getAll());
+        req.getRequestDispatcher("/WEB-INF/view/product/all.jsp").forward(req, resp);
     }
 }
