+++ src/main/java/com/internet/shop/controller/ShoppingCartController.java
@@ -1,4 +1,4 @@
-package com.internet.shop.controller.cart;
+package com.internet.shop.controller;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.model.ShoppingCart;
