+++ src/main/java/com/internet/shop/controller/UsersController.java
@@ -1,8 +1,8 @@
-package com.internet.shop.controller.admin;
+package com.internet.shop.controller;
 
 import com.internet.shop.lib.Injector;
-import com.internet.shop.model.Order;
-import com.internet.shop.service.OrderService;
+import com.internet.shop.model.User;
+import com.internet.shop.service.UserService;
 import java.io.IOException;
 import java.util.List;
 import javax.servlet.ServletException;
@@ -11,17 +11,16 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/admin/orders")
-public class AdminOrdersController extends HttpServlet {
+@WebServlet("/user/all")
+public class UsersController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("com.internet.shop");
-    private OrderService orderService = (OrderService) injector
-            .getInstance(OrderService.class);
+    private UserService userService = (UserService) injector.getInstance(UserService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Order> allOrders = orderService.getAll();
-        req.setAttribute("orders", allOrders);
-        req.getRequestDispatcher("/WEB-INF/view/admin/orders.jsp").forward(req, resp);
+        List<User> allUsers = userService.getAll();
+        req.setAttribute("users", allUsers);
+        req.getRequestDispatcher("/WEB-INF/view/user/all.jsp").forward(req, resp);
     }
 }
