--- src/main/java/com/internet/shop/controller/admin/AdminController.java
@@ -1,17 +1,0 @@
-package com.internet.shop.controller.admin;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/admin/main")
-public class AdminController extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        req.getRequestDispatcher("/WEB-INF/view/admin/main.jsp").forward(req, resp);
-    }
-}
