--- src/main/java/com/internet/shop/controller/cart/DeleteProductFromShoppingCartController.java
@@ -1,32 +1,0 @@
-package com.internet.shop.controller.cart;
-
-import com.internet.shop.lib.Injector;
-import com.internet.shop.model.Product;
-import com.internet.shop.model.ShoppingCart;
-import com.internet.shop.service.ProductService;
-import com.internet.shop.service.ShoppingCartService;
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/shopping-cart/delete")
-public class DeleteProductFromShoppingCartController extends HttpServlet {
-    private static final Long USER_ID = 1L;
-    private static final Injector injector = Injector.getInstance("com.internet.shop");
-    private ShoppingCartService shoppingCartService = (ShoppingCartService) injector
-            .getInstance(ShoppingCartService.class);
-    private ProductService productService = (ProductService) injector
-            .getInstance(ProductService.class);
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        Product product = productService.get(Long.parseLong(req.getParameter("id")));
-        ShoppingCart shoppingCart = shoppingCartService.getByUserId(USER_ID);
-        shoppingCartService.deleteProduct(shoppingCart, product);
-        resp.sendRedirect(req.getContextPath() + "/shopping-cart/products");
-    }
-}
