--- src/main/java/com/internet/shop/controller/order/DeleteOrderController.java
@@ -1,24 +1,0 @@
-package com.internet.shop.controller.order;
-
-import com.internet.shop.lib.Injector;
-import com.internet.shop.service.OrderService;
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/order/delete")
-public class DeleteOrderController extends HttpServlet {
-    private static final Injector injector = Injector.getInstance("com.internet.shop");
-    private OrderService orderService = (OrderService) injector
-            .getInstance(OrderService.class);
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        orderService.delete(Long.valueOf(req.getParameter("id")));
-        resp.sendRedirect(req.getContextPath() + "/admin/orders");
-    }
-}
