--- src/main/java/com/internet/shop/controller/product/DeleteProductController.java
@@ -1,24 +1,0 @@
-package com.internet.shop.controller.product;
-
-import com.internet.shop.lib.Injector;
-import com.internet.shop.service.ProductService;
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/product/delete")
-public class DeleteProductController extends HttpServlet {
-    private static final Injector injector = Injector.getInstance("com.internet.shop");
-    private ProductService productService = (ProductService) injector
-            .getInstance(ProductService.class);
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        productService.delete(Long.parseLong(req.getParameter("id")));
-        resp.sendRedirect(req.getContextPath() + "/admin/products");
-    }
-}
