--- src/main/java/com/internet/shop/controller/product/ProductsController.java
@@ -1,24 +1,0 @@
-package com.internet.shop.controller.product;
-
-import com.internet.shop.lib.Injector;
-import com.internet.shop.service.ProductService;
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/product/all")
-public class ProductsController extends HttpServlet {
-    private static final Injector injector = Injector.getInstance("com.internet.shop");
-    private ProductService productService = (ProductService) injector
-            .getInstance(ProductService.class);
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        req.setAttribute("products", productService.getAll());
-        req.getRequestDispatcher("/WEB-INF/view/product/all.jsp").forward(req, resp);
-    }
-}
