+++ src/main/webapp/WEB-INF/view/data.jsp
@@ -2,10 +2,9 @@
 <html>
 <head>
     <title>Inject Data</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
 </head>
 <body>
 <h1>Data was added to Storage!</h1>
-<a class="btn-success btn-lg" href="${pageContext.request.contextPath}/">Go to the main page</a>
+<a href="${pageContext.request.contextPath}/">Go to the main page</a>
 </body>
 </html>
