+++ src/main/webapp/WEB-INF/view/user/all.jsp
@@ -3,12 +3,10 @@
 <html>
 <head>
     <title>All users</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
 </head>
 <body>
-<div class="container">
 <h1>All users page:</h1>
-<table class="table table-bordered">
+<table border="1">
     <tr>
         <th>User ID</th>
         <th>Name</th>
@@ -22,13 +20,12 @@
                 <c:out value="${user.name}"/>
             </td>
             <td>
-                <a class="btn-danger btn-lg" href="${pageContext.request.contextPath}/user/delete?id=${user.id}">Delete</a>
+                <a href="${pageContext.request.contextPath}/user/delete?id=${user.id}">Delete</a>
             </td>
         </tr>
     </c:forEach>
 </table>
 <br/>
-<a class="btn-info btn-lg" href="${pageContext.request.contextPath}/">Go to the main page</a>
-</div>
+<a href="${pageContext.request.contextPath}/">Go to the main page</a>
 </body>
 </html>
